Reconfigure shadcn to not use CSS variables
