+++ src/components/Navbar.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link"
 import Image from "next/image"
-import { Button } from "@/components/ui/button"
+import { Button } from "@/ui/button"
 import Form from "next/form"
 import { signIn } from "@/auth"
