+++ tsconfig.json
@@ -10,8 +10,7 @@
     "moduleResolution": "bundler",
     "noEmit": true,
     "paths": {
-      "@/*": ["./src/*"],
-      "@/ui/*": ["./src/components/ui/*"]
+      "@/*": ["./src/*"]
     },
     "plugins": [
       {
